refactor: move selector from App to LoginContainer
